website/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cross-web==0.4.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-nested-admin==4.1.4
django-prometheus==2.4.0
django-schema-viewer==0.5.3
django-widget-tweaks==1.5.0
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.33.0
googleapis-common-protos==1.70.0
graphql-core==3.2.7
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
jellyfish==1.2.0
jiter==0.11.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.3
langchain-core==1.2.7
langchain-openai==1.0.1
langchain-text-splitters==0.3.11
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.2
langsmith==0.4.27
libcst==1.8.6
markdown-it-py==4.0.0
mdurl==0.1.2
networkx==3.5
nltk==3.9.1
numpy==2.2.6
openai==2.6.1
orjson==3.11.3
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.0
pillow==11.3.0
prometheus_client==0.22.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-env==1.0.0
python-monkey-business==1.1.0
python-multipart==0.0.21
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
PyYAML-ft==8.0.0
rake-nltk==1.0.6
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
segtok==1.5.11
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.52.1
strawberry-graphql==0.288.4
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.12.0
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.13.0
uvicorn==0.40.0
websockets==15.0.1
whitenoise==6.9.0
xxhash==3.6.0
yake==0.4.8
zstandard==0.24.0
